Add optional syslog/journald logging
