+++ crates/cli/src/lib.rs
@@ -43,6 +43,34 @@ pub fn version_string() -> String {
     format!("rsync {ver}")
 }
 
+#[allow(clippy::vec_init_then_push)]
+pub fn version_banner() -> String {
+    #[allow(unused_mut)]
+    let mut features: Vec<&str> = Vec::new();
+    #[cfg(feature = "xattr")]
+    features.push("xattr");
+    #[cfg(feature = "acl")]
+    features.push("acl");
+    let features = if features.is_empty() {
+        "none".to_string()
+    } else {
+        features.join(", ")
+    };
+    let protocols = SUPPORTED_PROTOCOLS
+        .iter()
+        .map(|p| p.to_string())
+        .collect::<Vec<_>>()
+        .join(", ");
+    let upstream = option_env!("UPSTREAM_VERSION").unwrap_or("unknown");
+    format!(
+        "oc-rsync {} (rsync {})\nProtocols: {}\nFeatures: {}\n",
+        env!("CARGO_PKG_VERSION"),
+        upstream,
+        protocols,
+        features,
+    )
+}
+
 fn parse_filters(s: &str, from0: bool) -> std::result::Result<Vec<Rule>, filters::ParseError> {
     let mut v = HashSet::new();
     parse_with_options(s, from0, &mut v, 0)
@@ -117,48 +145,6 @@ fn parse_bool(s: &str) -> std::result::Result<bool, String> {
         _ => Err("invalid boolean".to_string()),
     }
 }
-
-pub fn version_string() -> String {
-    format!(
-        "oc-rsync {} (rsync {})\n",
-        env!("CARGO_PKG_VERSION"),
-        env!("UPSTREAM_VERSION"),
-    )
-}
-
-#[allow(clippy::vec_init_then_push)]
-pub fn version_banner() -> String {
-    #[allow(unused_mut)]
-    let mut features: Vec<&str> = Vec::new();
-    #[cfg(feature = "xattr")]
-    features.push("xattr");
-    #[cfg(feature = "acl")]
-    features.push("acl");
-    let features = if features.is_empty() {
-        "none".to_string()
-    } else {
-        features.join(", ")
-    };
-    let protocols = SUPPORTED_PROTOCOLS
-        .iter()
-        .map(|p| p.to_string())
-        .collect::<Vec<_>>()
-        .join(", ");
-    let upstream = option_env!("UPSTREAM_VERSION").unwrap_or("unknown");
-    format!(
-        version_string(),
-        "oc-rsync {} (rsync {})\nProtocols: {}\nFeatures: {}\n",
-        env!("CARGO_PKG_VERSION"),
-        upstream,
-        protocols,
-        features,
-    )
-}
-
-pub fn version_string() -> String {
-    version_banner()
-}
-
 pub fn parse_logging_flags(matches: &ArgMatches) -> (Vec<InfoFlag>, Vec<DebugFlag>) {
     let mut info: Vec<InfoFlag> = matches
         .get_many::<InfoFlag>("info")
@@ -172,6 +158,33 @@ pub fn parse_logging_flags(matches: &ArgMatches) -> (Vec<InfoFlag>, Vec<DebugFla
         .map(|v| v.copied().collect())
         .unwrap_or_default();
     (info, debug)
+}
+
+fn init_logging(matches: &ArgMatches) {
+    let verbose = matches.get_count("verbose");
+    let quiet = matches.get_flag("quiet");
+    let log_format = *matches
+        .get_one::<LogFormat>("log_format")
+        .unwrap_or(&LogFormat::Text);
+    let log_file = matches.get_one::<PathBuf>("client-log-file").cloned();
+    let log_file_fmt = matches.get_one::<String>("client-log-file-format").cloned();
+    let syslog = matches.get_flag("syslog");
+    let journald = matches.get_flag("journald");
+    let (mut info, mut debug) = parse_logging_flags(matches);
+    if quiet {
+        info.clear();
+        debug.clear();
+    }
+    logging::init(
+        log_format,
+        verbose,
+        &info,
+        &debug,
+        quiet,
+        log_file.map(|p| (p, log_file_fmt)),
+        syslog,
+        journald,
+    );
 }
 
 fn locale_charset() -> Option<String> {
@@ -900,6 +913,7 @@ struct ProbeOpts {
 }
 
 pub fn run(matches: &clap::ArgMatches) -> Result<()> {
+    init_logging(matches);
     let opts =
         ClientOpts::from_arg_matches(matches).map_err(|e| EngineError::Other(e.to_string()))?;
     if opts.daemon.daemon {
