+++ crates/logging/Cargo.toml
@@ -11,3 +11,8 @@ clap = { version = "4", features = ["derive"] }
 [dev-dependencies]
 tempfile = "3"
 
+[features]
+default = ["syslog", "journald"]
+syslog = []
+journald = []
+
